+++ src/services/obsidian/plugin-manager.ts
@@ -1,4 +1,3 @@
-import { OuterSetting } from "src/globals/plugin-setting";
 import { ChinesePatch } from "src/integrations/languages/chinese-patch";
 import { OmnisearchIntegration } from "src/integrations/omnisearch";
 import { AssetsProvider } from "src/utils/web/assets-provider";
@@ -9,17 +8,17 @@ import { CommandRegistry } from "./command-registry";
 import { SettingManager } from "./setting-manager";
 import { DataManager } from "./user-data/data-manager";
 import { ViewRegistry } from "./view-registry";
+import { FloatingWindowManager } from "src/ui/floating-window";
 
 @singleton()
 export class PluginManager {
-	private commandRegistry = getInstance(CommandRegistry);
 	// private readonly obFileUtil = getInstance(Vault).adapter as FileSystemAdapter;
 
 	async onload() {
 		await getInstance(SettingManager).initAsync();
 		getInstance(ViewRegistry).init();
 
-		this.commandRegistry.addCommandsWithoutDependency();
+		getInstance(CommandRegistry).addCommandsWithoutDependency();
 
 		await getInstance(AssetsProvider).initAsync();
 		await getInstance(ChinesePatch).initAsync();
@@ -29,12 +28,19 @@ export class PluginManager {
 	async onLayoutReady() {
 		await getInstance(DataManager).initAsync();
 		await getInstance(OmnisearchIntegration).initAsync();
-		this.commandRegistry.addInVaultLexicalCommands();
-		this.commandRegistry.addDevCommands();
+
+		const commandRegistry = getInstance(CommandRegistry);
+		commandRegistry.addInVaultLexicalCommands();
+		commandRegistry.addDevCommands();
 	}
 
 	// should be called in CleverSearch.onunload()
 	onunload() {
 		getInstance(DataManager).onunload();
+		getInstance(FloatingWindowManager).onunload();
+	}
+
+	onAppQuit() {
+		// getInstance(SettingManager).saveSettings();
 	}
 }